fp issue (it might be over bros)

+++ src/utils.c
@@ -2,7 +2,7 @@
 #include "utils.h"
 
 
-Interval interval_new(double min, double max) {
+Interval interval_new(float min, float max) {
   Interval interval = {.min=min, .max=max};
   return interval;
 }
@@ -10,7 +10,7 @@ float interval_sizeof(Interval interval) {
   return interval.max - interval.min;
 }
 
-float interval_clamp(Interval interval, double num) {
+float interval_clamp(Interval interval, float num) {
   if (num < interval.min) {
     return interval.min;
   } 
@@ -20,6 +20,6 @@ float interval_clamp(Interval interval, double num) {
   return num;
 }
 
-bool interval_surrounds(Interval interval, double x) {
+bool interval_surrounds(Interval interval, float x) {
   return interval.min < x && x < interval.max;
 }
